secp256k1: Move field val set hex to test file.
This moves the SetHex method on the field value type to the associated
test file so it is only avilable during the tests which is its intended
purpose.  It is not constant time and therefore shouldn't be available
to callers.  This is in preparation for exporting the type.
