+++ dcrec/secp256k1/curve.go
@@ -6,6 +6,7 @@
 package secp256k1
 
 import (
+	"encoding/hex"
 	"math/big"
 )
 
@@ -19,6 +20,18 @@ import (
 // (x, y) position on the curve, the Jacobian coordinates are (x1, y1, z1)
 // where x = x1/z1^2 and y = y1/z1^3.
 
+// hexToFieldVal converts the passed hex string into a fieldVal and will panic
+// if there is an error.  This is only provided for the hard-coded constants so
+// errors in the source code can be detected. It will only (and must only) be
+// called with hard-coded values.
+func hexToFieldVal(s string) *fieldVal {
+	b, err := hex.DecodeString(s)
+	if err != nil {
+		panic("invalid hex in source file: " + s)
+	}
+	return new(fieldVal).SetByteSlice(b)
+}
+
 var (
 	// Next 6 constants are from Hal Finney's bitcointalk.org post:
 	// https://bitcointalk.org/index.php?topic=3238.msg45565#msg45565
@@ -27,7 +40,7 @@ var (
 	// They have also been independently derived from the code in the
 	// EndomorphismVectors function in genstatics.go.
 	endomorphismLambda = fromHex("5363ad4cc05c30e0a5261c028812645a122e22ea20816678df02967c1b23bd72")
-	endomorphismBeta   = new(fieldVal).SetHex("7ae96a2b657c07106e64479eac3434e99cf0497512f58995c1396c28719501ee")
+	endomorphismBeta   = hexToFieldVal("7ae96a2b657c07106e64479eac3434e99cf0497512f58995c1396c28719501ee")
 	endomorphismA1     = fromHex("3086d221a7d46bcde86c90e49284eb15")
 	endomorphismB1     = fromHex("-e4437ed6010e88286f547fa90abfe4c3")
 	endomorphismA2     = fromHex("114ca50f7a8e2f3f657c1108d9d44cfd8")
@@ -37,7 +50,7 @@ var (
 	// seem to be about 8% slower in practice.
 	//
 	// endomorphismLambda = fromHex("AC9C52B33FA3CF1F5AD9E3FD77ED9BA4A880B9FC8EC739C2E0CFC810B51283CE")
-	// endomorphismBeta = new(fieldVal).SetHex("851695D49A83F8EF919BB86153CBCB16630FB68AED0A766A3EC693D68E6AFA40")
+	// endomorphismBeta = hexToFieldVal("851695D49A83F8EF919BB86153CBCB16630FB68AED0A766A3EC693D68E6AFA40")
 	// endomorphismA1 = fromHex("E4437ED6010E88286F547FA90ABFE4C3")
 	// endomorphismB1 = fromHex("-3086D221A7D46BCDE86C90E49284EB15")
 	// endomorphismA2 = fromHex("3086D221A7D46BCDE86C90E49284EB15")
